refactor(explorer): update Explorer.tsx to import from extracted files (#190)
Replace 1759-line god component with a slim orchestrator that imports
sub-components from components/explorer/. Re-exports FilesPanel and
types for backward compatibility.
